Add an example of sum operation usage.
With a single expression as its operand, if the expression resolves to
an array, $sum traverses into the array to operate on the numerical
elements of the array to return a single value.

With a list of expressions as its operand, if any of the expressions
resolves to an array, $sum does not traverse into the array but instead
treats the array as a non-numerical value.