Add an example of stream.Readable behaviour.
If a Readable is switched into flowing mode and there are no consumers
available to handle the data, that data will be lost. This can occur,
for instance, when the readable.resume() method is called without a
listener attached to the 'data' event, or when a 'data' event handler
is removed from the stream.